Ability to reset a nodes run_list from a node attr
We use chef to perform many tasks that are generally performed in
a late_command.  Once a node has performed the pre-requisite tasks
we want to reboot, and assign the node it's real worker roles.

This patch allows us to do so.  We pre-calculate and set node attributes
up front, prior to checking into chef.